More cert serial BigInteger/Long cleanup.
Missed some Long's that probably should go back to BigIntegers.

+++ proxy/src/main/java/org/fedoraproject/candlepin/service/EntitlementCertServiceAdapter.java
@@ -15,6 +15,7 @@
 package org.fedoraproject.candlepin.service;
 
 import java.io.IOException;
+import java.math.BigInteger;
 import java.security.GeneralSecurityException;
 import java.util.Date;
 import java.util.List;
@@ -47,7 +48,7 @@ public interface EntitlementCertServiceAdapter {
      */
     EntitlementCertificate generateEntitlementCert(Consumer consumer,
         Entitlement entitlement, Subscription sub, Product product, Date endDate, 
-        Long serialNumber) throws GeneralSecurityException, IOException;
+        BigInteger serialNumber) throws GeneralSecurityException, IOException;
 
     /**
      * Return a list of all entitlement certificates for a given consumer.
